+++ src/main/java/com/meli/socialmeli/repositories/IUserRepository.java
@@ -10,8 +10,5 @@ import com.meli.socialmeli.entities.User;
 public interface IUserRepository {
     List<User> findAll();
     User finById(int id);
-    User getUserById(int userId);
-    User updateUser(User user);
 
-    List<User> getFollowedUsers(int userId);
 }
