+++ src/main/java/com/meli/socialmeli/services/impl/UserServiceImpl.java
@@ -1,24 +1,23 @@
 package com.meli.socialmeli.services.impl;
 
-import com.meli.socialmeli.dtos.UserUnfollowDto;
+import com.meli.socialmeli.dtos.response.*;
 import com.meli.socialmeli.entities.User;
+
 import com.meli.socialmeli.repositories.IUserRepository;
-import com.meli.socialmeli.repositories.impl.UserRepositoryImpl;
+
 import com.meli.socialmeli.dtos.MessageDto;
-import com.meli.socialmeli.dtos.response.FollowersCountDto;
-import com.meli.socialmeli.dtos.response.UserResponseDto;
+
 import com.meli.socialmeli.exceptions.custom.BadRequestException;
-import com.meli.socialmeli.dtos.response.UserFollowersDTO;
-import com.meli.socialmeli.dtos.response.UserInfoDTO;
+
 import com.meli.socialmeli.exceptions.custom.NotFoundException;
+
 import com.meli.socialmeli.services.IUserService;
+
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.List;
-
 import java.util.Optional;
 
 @Service
@@ -115,4 +114,34 @@ public class UserServiceImpl implements IUserService {
         followersCountDto.setFollowers_count(user.getFollowers().size());
         return followersCountDto;
     }
+
+    @Override
+    public UserFollowersDTO findFollowersById(int id) {
+        Optional<User> userFound = Optional.ofNullable(userRepository.finById(id));
+        if (userFound.isEmpty()) throw new NotFoundException("There is no user with the id: " + id);
+
+        List<UserInfoDTO> followers = userFound.get().getFollowers()
+                .stream()
+                .map( f -> new UserInfoDTO(f.getUser_id(), f.getUser_name() ) )
+                .toList();
+
+        return new UserFollowersDTO(userFound.get().getUser_id(), userFound.get().getUser_name(), followers);
+    }
+
+    @Override
+    public UserUnfollowDto unfollowUser(int userId, int userIdToUnfollow) {
+        User user = this.userRepository.finById(userId);
+        if (user == null) {
+            throw new NotFoundException("User with id " + userId + " not found");
+        }
+        List<User> followedUsers = user.getFollowed();
+        boolean removed = followedUsers.removeIf((followedUser) -> followedUser.getUser_id() == userIdToUnfollow);
+        if (!removed) {
+            throw new NotFoundException("Followed user with id " + userIdToUnfollow + " not found");
+        }
+        user.setFollowed(followedUsers);
+
+        return new UserUnfollowDto(userId, userIdToUnfollow);
+    }
+
 }
